==> exploratory_notebooks/main/01_apply_rir.ipynb <==
# Descripiton

## Setup

# Need newer torch to use all augmentation functions
%pip install torch==2.0.* torchaudio

## Reload to get installed packages

import os
os._exit(00)

## Imports

import glob
import os
import multiprocessing
import random

from itertools import repeat, product

from collections import namedtuple
from csv import writer

import torch
import torchaudio
import pandas as pd
import tqdm

import tqdm.contrib.itertools as tqdm_itertools
from IPython.display import Audio
from datasets import Audio as DSAudio, load_dataset

tqdm_chain = tqdm_itertools.itertools.chain

### Constants and parameters

SOURCE_SAMPLING_RATE = 44100
MODEL_SAMPLING_RATE = 32000

DATA_TAG = 'data_8c86715'

## Hyperparameters

## Base entities

### Datapoint struct

DataPoint = namedtuple('DataPoint', ['path', 'label', 'name'])

### Paths

base_dir = os.path.abspath(os.path.join(os.getcwd(), '../../'))
data_source_base = os.path.join(base_dir, 'gdsc_data/data_source')
data_processed_base = os.path.join(base_dir, 'gdsc_data/data_processed')

train_data_source_path = os.path.join(data_source_base, 'train/') 
test_data_source_path = os.path.join(data_source_base, 'test/') 

rir_library_path = os.path.join(base_dir, 'audio_data/rir_library_flac_sampled/')

### Helper functions

def resample_source_wav(data: DataPoint, resampler):
    wav, _ = torchaudio.load(data.path)
    return resampler(wav)    


def iter_datapoint_from(source_data_path):
    metadata_file = os.path.join(source_data_path, 'metadata.csv')
    df_audio = pd.read_csv(metadata_file)
    for fname, label in df_audio.itertuples(index=False):
        full_path = os.path.join(source_data_path, fname)
        yield DataPoint(path=full_path, label=label, name=fname)
        
        
def rir_path_to_name_path(rir_path):
    name, ext = os.path.splitext(os.path.basename(rir_path))
    return name, rir_path
            

## Worker's init and main functions

def worker_init(worker_exec, source_dir, processed_dir):
    import torch
    import torchaudio
    torch.set_num_threads(1)
    
    worker_exec.torch = torch
    worker_exec.torchaudio = torchaudio
    worker_exec.source_dir = source_dir
    worker_exec.processed_dir = processed_dir
    worker_exec.resampler = torchaudio.transforms.Resample(SOURCE_SAMPLING_RATE, MODEL_SAMPLING_RATE)
    
    
def worker_main(rir_name, rir_path, data_point):
    # Load initalized libraries in a process
    torch = worker_main.torch
    torchaudio =  worker_main.torchaudio
    processed_dir = worker_main.processed_dir
    source_dir = worker_main.source_dir
    resampler = worker_main.resampler
    
    # convert to 32kHZ
    wav_resampled = resample_source_wav(data_point, resampler)
    
    # get the room IR, normalize it, then apply
    rir, _ = torchaudio.load(rir_path)
    rir = rir / torch.norm(rir, p=2)
    wav_resampled = torchaudio.functional.fftconvolve(wav_resampled, rir)
    
    # prepare for output
    source_path = data_point.path
    dp_label = data_point.label    
    dp_name = data_point.name
    source_name, source_ext = os.path.splitext(dp_name)
    
    # This will recreate train/ subfolder as well
    dest_name = f'{source_name}@conv_{rir_name}.flac'
    dest_relpath = os.path.dirname(os.path.relpath(source_path, source_dir))
    dest_fpath = os.path.join(processed_dir, dest_relpath, dest_name)  
    
    os.makedirs(os.path.dirname(dest_fpath), exist_ok=True)
    
    torchaudio.save(
        dest_fpath,
        wav_resampled,
        sample_rate=MODEL_SAMPLING_RATE,
        format="flac", 
        bits_per_sample=24)

    return dest_fpath, dest_name, dp_label 


def worker_starmap(args):
    return worker_main(*args)

## Scheduler

RIR_SAMPLE_SIZE = 20

rir_library = list(map(rir_path_to_name_path, glob.glob(os.path.join(rir_library_path, '**/*.flac'), recursive=True)))
source_library = iter_datapoint_from(train_data_source_path)

all_combinations = product(random.sample(rir_library, RIR_SAMPLE_SIZE), source_library)

source_library_length = iter_datapoint_from(train_data_source_path)
all_combinations_length = 0
for i, _ in enumerate(product(range(0, RIR_SAMPLE_SIZE), source_library_length), 1):
    all_combinations_length = i

work_generator = ((rir_name, rir_path, dp) for (rir_name, rir_path), dp in all_combinations)

print(f"Total possible rir+sound combinations {all_combinations_length}")

# if raises exception, make sure you delete the folder with all the data first
dest_path = os.path.join(data_processed_base, DATA_TAG, '01_apply_ir_function/')
os.stat(dest_path)

print(f"Reading files from '{train_data_source_path}'")
print(f"Going to write into '{dest_path}'")

n_cpu = multiprocessing.cpu_count()
pool = multiprocessing.Pool(processes=n_cpu, initializer=worker_init, initargs=(worker_main, data_source_base, dest_path))

rows = []
for _, dest_fname, label in tqdm.tqdm(pool.imap_unordered(worker_starmap, work_generator), total=all_combinations_length):
    rows.append((dest_fname, label))

del pool

df = pd.DataFrame(rows, columns=['file_name', 'label'])
df.to_csv(os.path.join(dest_path, 'metadata.csv'), index=False)

## Change ds a bit - map between original and all "variants" available as a comma separated list

df = pd.read_csv(os.path.join(dest_path, 'metadata.csv'))

def extract_original_filename(file_name):
    parts = file_name.split("@")
    if len(parts) == 0:
        return file_name
    
    return f"{parts[0]}.wav"

df["original_filename"] = df["file_name"].map(extract_original_filename)

df_grouped = df.groupby("original_filename").aggregate(
    {
        "file_name": lambda x: ", ".join(x),
        "label": min
    }
).reset_index().rename({"file_name": "file_list"})

df_grouped.to_csv(os.path.join(dest_path, 'metadata_grouped_train.csv'), index=False)

df_grouped["file_name"].map(lambda x: x.split(", "))[0]

## Merge validation with train
We need more data for training, and we'll opt for a tiny validation dataset.
Then, the downsized validation dataset will undergo all the IR again, to make it bigger. 

We'll only take one sample per species for validation ds

val_data_source_path = os.path.join(data_source_base, 'val/')

df_grouped

## Legacy processing (TO BE REMOVED)

def split_nsec_chunks(data_point: DataPoint, sampling_rate, n_sec=10):
    dpath = data_point.path
    data, _ = torchaudio.load(dpath)
    name = data_point.name
    
    
    if data.shape[1] <= sampling_rate * n_sec:
        raw_name, ext = os.path.splitext(name)
        yield data, f"{raw_name}_1{ext}", data_point.label
        
        return
    
    for i, chunk in enumerate(torch.split(data, n_sec * sampling_rate, dim=1), 1):
        raw_name, ext = os.path.splitext(name)
        yield chunk, f"{raw_name}_{i}{ext}", data_point.label  
        
        
resampler = torchaudio.transforms.Resample(SOURCE_SAMPLING_RATE, MODEL_SAMPLING_RATE)   
def resample_wavs(data, fname, processed_dir, sr=MODEL_SAMPLING_RATE, sampler=resampler):
        resampled_wav = resampler(data)
        torchaudio.save(os.path.join(processed_dir, fname), resampled_wav, sample_rate=MODEL_SAMPLING_RATE)

        
def make_data_point(data_path):
    """ Assumes that metadata.csv lies along with the .wav files
    """
    metadata = os.path.join(data_path, 'metadata.csv')
    df_audio = pd.read_csv(metadata)
    
    for fname, label in df_audio.itertuples(index=False):
        full_path = os.path.join(data_path, fname)
        yield DataPoint(path=full_path, label=label, name=fname)


### Create new directories

os.makedirs(os.path.join(data_processed, 'train/'), exist_ok=True)
os.makedirs(os.path.join(data_processed, 'test/'), exist_ok=True)
os.makedirs(os.path.join(data_processed, 'val/'), exist_ok=True)

## Split files and append to metadata file

`tqdm` is really buggy. it seems to work for the first run, but doesn't run for a second iteration, unless variable names change. Therefore outcommented 

def preprocess_files(store_at, source_metadata_path):
    '''
    Moved to function. Need to reconsider how test dataset gets handled.
    Also REALLY want to get tdqm to work....DONE
    '''

    #(Re)create CSV file with headers
    with open(store_at + 'metadata.csv', 'w') as f_object:
        writer_object = writer(f_object)
        writer_object.writerow(['file_name', 'label'])
        f_object.close()
    
    # GSK: Would love to have this snippet running, but too inconsistent.
    #main_iterable_v234 = chain.from_iterable(map(split_nsec_chunks, make_data_point(_data, max_num=_max_num), repeat([SOURCE_SAMPLING_RATE])))
    #total_len = sum(1 for _ in main_iterable_v234)
    #print(total_len)
    #for data, fname, label in tqdm.tqdm(main_iterable_v234):#, total = total_len):
    for data, fname, label in tqdm.tqdm((map(split_nsec_chunks, make_data_point(source_metadata_path), repeat(SOURCE_SAMPLING_RATE)))):
        # Resample and save .wav
        resample_wavs(data, fname, store_at)
                
        # Write to CSV
        with open(store_at + 'metadata.csv', 'a') as f_object:
            writer_object = writer(f_object)
            writer_object.writerow([fname, label])
            f_object.close()

preprocess_files(store_at = os.path.join(data_processed, 'train/'), source_metadata_path = train_data_source_path)
preprocess_files(store_at = os.path.join(data_processed, 'val/'), source_metadata_path = val_data_source_path)

## Split the test dataset
Could be added to other methods, but likely requires lots of changes, and this only needs to be done once.

n_sec=10

# Read in the file locations
metadata = os.path.join(test_data_path, 'metadata.csv')
df_audio = pd.read_csv(metadata)

#(Re)create CSV file with headers
with open(os.path.join(data_processed, 'test/') + 'metadata.csv', 'w') as f_object:
    writer_object = writer(f_object)
    writer_object.writerow(['file_name'])
    f_object.close()

cnt = 0
# Go through each file
for fname in tqdm.tqdm(df_audio.itertuples(index=False)):
    #print(cnt)
    cnt = cnt +1 
    # Extract the information
    fname = fname[0]
    full_path = os.path.join(test_data_path, fname)
    data, _ = torchaudio.load(full_path)
    
    # Chunk it up
    for i, chunk in enumerate(torch.split(data, n_sec * SOURCE_SAMPLING_RATE, dim=1), 1):
        raw_name, ext = os.path.splitext(fname)
        outfile_name = f"{raw_name}_{i}{ext}"
        resample_wavs(chunk, outfile_name, os.path.join(data_processed, 'test/'), sr=MODEL_SAMPLING_RATE, sampler=resampler)    
        
        # Write to CSV
        with open(os.path.join(data_processed, 'test/') + 'metadata.csv', 'a') as f_object:
            writer_object = writer(f_object)
            writer_object.writerow([outfile_name])
            f_object.close()
            
#    if cnt > 5:
#        break

#Audio(full_path, rate = 22000)

#Audio(os.path.join(processed_dir, 'test/2_2.wav'), rate=32000)

# Old snippets I didn't want to delete just yet

Audio(list(split_nsec_chunks(DataPoint(props=train_data[1411]['audio'], label=train_data[1411]['label']), SOURCE_SAMPLING_RATE))[10][0], rate=SOURCE_SAMPLING_RATE)

### Sample to search for is 1411 - it's 1 min long

for i, data in enumerate(train_data):
    if os.path.basename(data['audio']['path']) == 'Roeselianaroeselii_XC751814-dat028-019_edit1.wav':
        print(i)

### Backup, because I'm about to break it with TQDM:

def preprocess_files(_processing_directory, _data, _max_num = 10):
    '''
    Moved to function. Need to reconsider how test dataset gets handled.
    Also REALLY want to get tdqm to work....
    '''
    # Dummy for development:
    print("----------------------------")
    print(f"processing up to {_max_num} files in {_processing_directory}")
    print("----------------------------")
    
    #(Re)create CSV file with headers
    with open(_processing_directory + 'metadata.csv', 'w') as f_object:
        writer_object = writer(f_object)
        writer_object.writerow(['filename', 'label'])
        f_object.close()

    # Loop through clips, save as files and append to metadata.csv
    cnt = 0
    
    #for data, fname, label in chain.from_iterable(map(split_nsec_chunks, make_data_point(_data, max_num=_max_num), repeat([SOURCE_SAMPLING_RATE]))):
    for data, fname, label in chain.from_iterable(map(split_nsec_chunks, make_data_point(_data, max_num=_max_num), repeat([SOURCE_SAMPLING_RATE]))):
        cnt = cnt+1
        
        # Resample and save .wav
        resample_wavs(data, fname, _processing_directory)
                
        # Write to CSV
        with open(_processing_directory + 'metadata.csv', 'a') as f_object:
            writer_object = writer(f_object)
            writer_object.writerow([fname, label])
            f_object.close()
            
        print(f'{cnt}: {label} - {fname}')
