==> exploratory_notebooks/main/11_fine_tune_passt.ipynb <==
%pip install -e 'src/hear21passt'

%pip install tensorboard

import os
os._exit(00)

import sys
import os
import datetime
import re

import numpy as np
import pandas as pd
import torch
import torchaudio
import tqdm

from operator import itemgetter
from hashlib import sha1

from torch.optim import AdamW, lr_scheduler
from torch.utils.data import DataLoader, default_collate
from torch.nn import functional as F
from torch.utils.tensorboard import SummaryWriter
from torch.utils.data import ChainDataset

from sklearn.metrics import f1_score

from IPython.display import Audio

from optim.lr_scheduler import get_scheduler_lambda
from utils.model import get_pretrained_passt_model
from utils.gdsc22_dataset import BuzzIterableDataset, BuzzAugmentedIterableDataset
from utils.dataset_augmentations import BuzzIterableTransformedDataset, MixUp1Level, MixUp2Level, RandomGain, Rolling, ColoredNoise

## 1. Initial configuration

torch.manual_seed(42)
np.random.seed(42)

project_dir = os.path.abspath(os.path.join(os.getcwd(), '../../'))
data_processed_dir = os.path.join(project_dir, 'gdsc_data/data_processed/')
data_source_dir = os.path.join(project_dir, 'gdsc_data/data_source/')

ts = datetime.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')

MODEL_NAME = f"gdsc22_passt_kd_NOEXP_lr_mismatch_size_t40_f4_epochs40_gain_roll_mixup_noise@{ts}.pt"
TORCH_SAVE_PATH = os.path.join(project_dir, f'checkpoints/{MODEL_NAME}')

TENSORBOARD_RUNS_PATH = os.path.join(project_dir, "torch_runlogs/")
TENSORBOARD_EXP_FOLDER = f"PaSST_kd_NOEXP_lr_mismatch_size_t40_f4_epochs40_gain_roll_mixup_noise@{ts}"
TENSORBOARD_EXP_COMMENT = "PaSST_s_kd_p16_128_ap486_NOEXP_lr_gain_roll_mixup_noise"

TRAIN_BATCH_SIZE = 32
VAL_BATCH_SIZE = 32

# fine tuning and ohe of real labels
CLASSIFIER_N_CLASSES = 66

DATA_TAG = "data_8c86715"

DEBUG_TRANSFORMATIONS = False

#final_dataset_path = find_final_dataset_path(data_processed_dir, DATA_TAG)
final_dataset_path_augmented = os.path.join(data_processed_dir, DATA_TAG, '01_apply_ir_function')
final_dataset_val_as_train_path_augmented = os.path.join(data_processed_dir, DATA_TAG, '03_apply_ir_function_on_validation')

final_dataset_path = data_source_dir
final_dataset_val_as_train_path = os.path.join(data_processed_dir, DATA_TAG, '02_resample_validation/')

train_dir_augmented = os.path.join(final_dataset_path_augmented, 'train/')
val_as_train_dir_augmented = os.path.join(final_dataset_val_as_train_path_augmented, 'train/')

train_dir = os.path.join(final_dataset_path, 'train/')
val_as_train_dir = os.path.join(final_dataset_val_as_train_path, 'train/')

val_dir = os.path.join(final_dataset_val_as_train_path, 'val/')

## 2. Dataloaders and batching

### 2.1 Debugging collate_fn

Used to check the random seed and batchability of the individual points 0:)

def debug_collate(batch):
    get_1st = itemgetter(0)
    files = ", ".join(map(get_1st, batch)).encode('utf-8')
    worker_info = torch.utils.data.get_worker_info()   
    print(f"Worker_id '{worker_info.id}' batch hash is '{sha1(files).hexdigest()}'") 
    
    batch = default_collate(batch)
    return batch

train_dataset = BuzzIterableDataset(
    train_dir,
    metadata_csv=os.path.join(train_dir, 'metadata.csv'), 
    shuffle=True,
    subsample=1.0,
    total_classes=CLASSIFIER_N_CLASSES
)
val_dataset = BuzzIterableDataset(
    val_as_train_dir,
    metadata_csv=os.path.join(val_as_train_dir, 'metadata.csv'), 
    subsample=1.0,
    total_classes=CLASSIFIER_N_CLASSES
)

train_dataset_transformed = BuzzIterableTransformedDataset(
    transformers=[RandomGain(debug=DEBUG_TRANSFORMATIONS), Rolling(debug=DEBUG_TRANSFORMATIONS), MixUp1Level(debug=DEBUG_TRANSFORMATIONS), ColoredNoise(debug=DEBUG_TRANSFORMATIONS)],
    buzz_iterable=train_dataset
)

# I had too much hope on the IR augmented dataset and took a wrong path...
# train_dataset_augmented = BuzzAugmentedIterableDataset(
#     train_dir_augmented, 
#     metadata_csv=os.path.join(final_dataset_path_augmented, 'metadata_grouped_train.csv'), 
#     shuffle=True,
#     subsample=0.2,
#     total_classes=CLASSIFIER_N_CLASSES
# )
# val_as_train_dataset_augmented = BuzzAugmentedIterableDataset(
#     val_as_train_dir_augmented, 
#     metadata_csv=os.path.join(final_dataset_val_as_train_path_augmented, 'metadata_grouped_train.csv'), 
#     shuffle=True,
#     subsample=0.2,
#     total_classes=CLASSIFIER_N_CLASSES
# )
    
train_dataloader = DataLoader(
    train_dataset_transformed, 
    batch_size=TRAIN_BATCH_SIZE, 
    pin_memory=True,
    num_workers=4,
    #collate_fn=debug_collate
)


# val_dataset = BuzzIterableDataset(
#     val_dir, 
#     metadata_csv=os.path.join(val_dir, 'metadata.csv'),
#     shuffle=False,
#     total_classes=CLASSIFIER_N_CLASSES,
# )
val_dataloader = DataLoader(val_dataset, batch_size=VAL_BATCH_SIZE)

## 3. Init pytorch objects

writer = SummaryWriter(os.path.join(TENSORBOARD_RUNS_PATH, TENSORBOARD_EXP_FOLDER), comment=TENSORBOARD_EXP_COMMENT)
device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')

%%sh
nvidia-smi

# Default stride for t and f is 10
# NOTE: FMAX is used in frequency augmentation. Every sample has a random subset of fmin and fmax applied to it...
model = get_pretrained_passt_model(mode="all", n_classes=66, s_patchout_t=40, s_patchout_f=4)

#model = get_pretrained_passt_model(arch='passt_s_swa_p16_128_ap476', mode="all", n_classes=66, s_patchout_t=10, s_patchout_f=3)
model.to(device)

adamw = AdamW(model.net.parameters(), lr=0.00001, weight_decay=0.0001)
lr_scheduler = torch.optim.lr_scheduler.LambdaLR(adamw, get_scheduler_lambda(warm_up_len=0))

scaler = torch.cuda.amp.GradScaler()

print(f"Initial Learning rate:{lr_scheduler.get_last_lr()[0]}")
for epoch in range(0, 40):
    y_tr_pred_f1 = []
    y_tr_true_f1 = []
    train_loss = []

    model.train()
    for _, _, X_batch, y in (progress := tqdm.tqdm(train_dataloader)):
        adamw.zero_grad()        
        
        X_batch = X_batch.to(device)
        y = y.to(device)
        X_batch_spectr = model.mel(X_batch).unsqueeze(1)
        
        with  torch.cuda.amp.autocast(): 
            y_hat, emb = model.net(X_batch_spectr)     
            
            loss = F.cross_entropy(y_hat, y, reduction="none")
            loss = loss.mean()
            
            train_loss.append(loss.detach().cpu())
            progress.set_description(f"TrainL {float(loss):.2f}")
            
        # Scales loss.  Calls backward() on scaled loss to create scaled gradients.
        # Backward passes under autocast are not recommended.
        # Backward ops run in the same dtype autocast chose for corresponding forward ops.
        scaler.scale(loss).backward()
        # scaler.step() first unscales the gradients of the optimizer's assigned params.
        # If these gradients do not contain infs or NaNs, optimizer.step() is then called,
        # otherwise, optimizer.step() is skipped.
        scaler.step(adamw)
        # Updates the scale for next iteration.
        scaler.update()
        
        prob = F.softmax(y_hat.detach(), dim=1)
        y_pred = torch.argmax(prob.detach(), dim=1)
        y_tr_pred_f1.append(y_pred.cpu())
        y_tr_true_f1.append(y.cpu().argmax(dim=1))

    f_score_train = f1_score(torch.cat(y_tr_true_f1).numpy(), torch.cat(y_tr_pred_f1).numpy(), average='macro')
    total_train_loss = torch.stack(train_loss, dim=0).mean()
    print(f"=" * 80)
    print("\r\n")
    print(f"Train Loss(epoch={epoch}): {total_train_loss}")
    print(f"F1 score(train) {f_score_train}")
    print("\r\n")
    print(f"=" * 80)
    print("\r\n")

    val_loss = []
    y_val_pred_f1 = []
    y_val_true_f1 = []
    model.eval()
    
    with torch.no_grad():
        for _, _, X_batch, y_val in (progress := tqdm.tqdm(val_dataloader)):
            X_batch = X_batch.to(device)
            y_val = y_val.to(device)
            X_batch_spectr = model.mel(X_batch).unsqueeze(1)

            with torch.cuda.amp.autocast():
                y_hat, emb = model.net(X_batch_spectr)
            
                _loss = F.cross_entropy(y_hat, y_val, reduction="none")
                _loss = _loss.mean()
                val_loss.append(_loss.detach().cpu())
            
            progress.set_description(f"ValL {float(_loss):.2f}")            

            prob = F.softmax(y_hat.detach(), dim=1)
            y_pred = torch.argmax(prob.detach(), dim=1)

            y_val_pred_f1.append(y_pred.cpu())
            y_val_true_f1.append(y_val.cpu().argmax(dim=1))
            
    f1_score_val = f1_score(torch.cat(y_val_true_f1).numpy(), torch.cat(y_val_pred_f1).numpy(), average='macro')
    total_val_loss = torch.stack(val_loss, dim=0).mean()
    print(f"=" * 80)
    print("\r\n")
    print(f"Val Loss: {total_val_loss}")
    print(f"F1 score(val) {f1_score_val}")
    print("\r\n")    
    print(f"=" * 80)
    print("\r\n")
       
    writer.add_scalar('Loss/train', total_train_loss, epoch)
    writer.add_scalar('Loss/val', total_val_loss, epoch)
    writer.add_scalar('F1/train', f_score_train, epoch)
    writer.add_scalar('F1/val', f1_score_val, epoch)
    writer.add_scalar('LR', lr_scheduler.get_last_lr()[0], epoch)

    lr_scheduler.step()
    print(f"New lr: {lr_scheduler.get_last_lr()[0]}")
       
    torch.save(
        {
            'epoch': epoch,
            'model_state_dict': model.state_dict(),
            'optimizer_state_dict': adamw.state_dict(),
        }, TORCH_SAVE_PATH)

model_eval = get_pretrained_passt_model(mode="all", n_classes=66, s_patchout_t=0, s_patchout_f=0)
state = torch.load(TORCH_SAVE_PATH)
model_eval.load_state_dict(state['model_state_dict'])

model_eval = model_eval.cuda()

    val_loss = 0.0
    ypred_f1 = []
    ytrue_f1 = []
    model_eval.eval()
    
    with torch.no_grad():
        for batch, y_val in tqdm.tqdm(val_dataloader):
            batch = batch.cuda()
            y_val = y_val.cuda()

            y_hat, emb = model_eval(batch)
            _loss = F.cross_entropy(y_hat, y_val, reduction="none")

            prob = F.softmax(y_hat, dim=1)
            y_pred = torch.argmax(prob, dim=1)

            ypred_f1.append(y_pred.cpu())
            ytrue_f1.append(y_val.cpu())

            val_loss += _loss.sum()

    f1_score_val = f1_score(torch.cat(ytrue_f1).numpy(), torch.cat(ypred_f1).numpy(), average='macro')

    print(f"Val Loss: {val_loss}")
    print(f"=" * 80)
    print(f"F1 score(val) {f1_score_val}")
    print(f"=" * 80)

%%sh
nvidia-smi


## Miscellaneous

### Glossary

SWA - Stochastic weight averaging

### Model stored weights
```
Downloading: "https://github.com/kkoutini/PaSST/releases/download/v0.0.1-audioset/passt-s-f128-p16-s10-ap.476-swa.pt" to /root/.cache/torch/hub/checkpoints/passt-s-f128-p16-s10-ap.476-swa.pt
```
