==> exploratory_notebooks/main/90_sample_rir_library.ipynb <==
import os
import re
import shutil
import random

## Downsample rir library, only use one of the "rir" files in a folder. Save it in a separate directory
base_dir = os.path.abspath(os.path.join(os.getcwd(), '../../'))
rir_library_path = os.path.join(base_dir, 'audio_data/rir_library_flac/')
rir_processed_library_path = os.path.join(base_dir, 'audio_data/rir_library_flac_sampled/')

for dirpath, dirnames, filenames in os.walk(rir_library_path):
    if re.match(r'.*\s\(sr\).*', dirpath):
        continue    
    elif not filenames:
        continue
     
    #copy all files
    if dirpath.find('LX48L') > 0 or dirpath.find('Synthetic') > 0:
        units = filenames
    # sample one
    else:
        units = [random.choice(filenames)]
        
    relpath_to_recreate = os.path.relpath(dirpath, rir_library_path)
    abspath_to_recreate = os.path.join(rir_processed_library_path, relpath_to_recreate)

    os.makedirs(abspath_to_recreate, exist_ok=True)            
    for unit in units:
        shutil.copy(os.path.join(dirpath, unit), os.path.join(abspath_to_recreate, unit))
